업로드 파일에 해당하는 gen_file row가 생성되도록

+++ src/main/java/com/ll/exam/app10/app/fileUpload/entity/GenFile.java
@@ -14,15 +14,27 @@ import javax.persistence.Entity;
 @SuperBuilder
 @ToString(callSuper = true)
 public class GenFile extends BaseEntity {
+
     private String relTypeCode;
-    private int relId;
+
+    private long relId;
+
     private String typeCode;
+
     private String type2Code;
+
     private String fileExtTypeCode;
+
     private String fileExtType2Code;
+
     private int fileSize;
+
     private int fileNo;
+
     private String fileExt;
+
     private String fileDir;
+
     private String originFileName;
+
 }
